- refactor: 사용 목적에 맞는 파일명으로 변경
