+++ src/App.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import VendingMachine from './components/VendingMachine';
 import Controls from './components/Controls';
-import {VendingMachineProvider} from './context/AppContext';
+import {VendingMachineProvider} from './context/VendingMachineContext';
 
 const App: React.FC = () => {
     return (
